Create Rock Paper Scissor Game.py
This programme uses conditional statements and functions to imitate the widely played game Rock, Paper, Scissors. What better approach, then, to master these important ideas? Additionally, you will be storing a set of legitimate answers in a Python list, which you may use to construct a sophisticated and Pythonic conditional expression.

This programme uses the random, os, and re modules from the standard library. It is one of several Python coding projects that imports other libraries.

This Python project idea invites the user to initiate action by passing in a character that represents rock, paper, or scissors. You can see the code for it below. Following the input string's evaluation, the conditional logic determines the winner.